requirements.txt
adversarial_robustness_toolbox==1.14.1
Jinja2==3.1.2
matplotlib==3.3.3
numpy==1.21.2
pandas==1.5.3
Pillow==8.0.1
scikit_learn==1.0.2
scipy==1.10.1
seaborn==0.11.2
tensorflow==2.8.0
torch==1.11.0
torchattacks==3.3.0
torchvision==0.15.0
tqdm==4.62.3
